refactor: consolidate installHelmChartNative logic into helm.go and remove helm_native.go
